+++ src/main/java/dio/estudos/gof/model/ClienteRepository.java
@@ -2,6 +2,6 @@ package dio.estudos.gof.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClienteRepository extends CrudRepository<Cliente, Long> {
+public interface ClienteRepository extends CrudRepository<Cliente, Integer> {
 
 }
